Repository Restructiring pt.2: Refactor database module structure. move TransactionsTable and related classes to new transactions.py, update imports across the codebase, and remove obsolete db_models.py.
